inital comment addressings
