+++ src/nft/utils/index.ts
@@ -8,4 +8,5 @@ export * from './isVideo'
 export * from './listNfts'
 export * from './putCommas'
 export * from './rarity'
-export * from './transaction'
+export * from './transactionResponse'
+export * from './updatedAssets'
