+++ src/nft/utils/updatedAssets.ts
@@ -9,3 +9,14 @@ export const updatedAssetPriceDifference = (asset: UpdatedGenieAsset) => {
 export const sortUpdatedAssets = (x: UpdatedGenieAsset, y: UpdatedGenieAsset) => {
   return updatedAssetPriceDifference(x).gt(updatedAssetPriceDifference(y)) ? -1 : 1
 }
+
+export const getTotalNftValue = (nfts: UpdatedGenieAsset[]): BigNumber => {
+  return (
+    nfts &&
+    nfts.reduce(
+      (ethTotal, nft) =>
+        ethTotal.add(BigNumber.from(nft.updatedPriceInfo ? nft.updatedPriceInfo.ETHPrice : nft.priceInfo.ETHPrice)),
+      BigNumber.from(0)
+    )
+  )
+}
